requirements.txt
Flask
google-genai
groq
mistralai
feedparser
requests
apscheduler
python-dotenv
dnspython
pydantic
pytest
